[pldmd]Use lowest data type possible-NumericSensor
Maximum possible value for sensor reading is 4 bytes in length.
Update numeric sensors to use lowest possible data size.

Tested:
Sensor values are exposed as expected on D-Bus
introspect xyz.openbmc_project.pldm /xyz/openbmc_project/sensors/temperature/outlet_temp_2
 NAME                                                  TYPE      SIGNATURE RESULT/VALUE FLAGS
 org.freedesktop.DBus.Introspectable                   interface -         -            -
 .Introspect                                           method    -         s            -
 org.freedesktop.DBus.Peer                             interface -         -            -
 .GetMachineId                                         method    -         s            -
 .Ping                                                 method    -         -            -
 org.freedesktop.DBus.Properties                       interface -         -            -
 .Get                                                  method    ss        v            -
 .GetAll                                               method    s         a{sv}        -
 .Set                                                  method    ssv       -            -
 .PropertiesChanged                                    signal    sa{sv}as  -            -
 xyz.openbmc_project.Sensor.Threshold.Critical         interface -         -            -
 .CriticalAlarmHigh                                    property  b         true         emits-change
 .CriticalAlarmLow                                     property  b         true         emits-change
 .CriticalHigh                                         property  d         0            emits-change
 .CriticalLow                                          property  d         0            emits-change
 xyz.openbmc_project.Sensor.Value                      interface -         -            -
 .MaxValue                                             property  d         254          emits-change
 .MinValue                                             property  d         0            emits-change
 .Value                                                property  d         0            emits-change
 xyz.openbmc_project.State.Decorator.Availability      interface -         -            -
 .Available                                            property  b         true         emits-change writable
 xyz.openbmc_project.State.Decorator.OperationalStatus interface -         -            -
 .Functional                                           property  b         true         emits-change

+++ pldmd/include/pdr_utils.hpp
@@ -28,18 +28,18 @@ namespace sensor
 
 /** @brief Calculate sensor value*/
 double calculateSensorValue(const pldm_numeric_sensor_value_pdr& pdr,
-                            const double& sensorReading);
+                            const float& sensorReading);
 
 /** @brief Fetch the sensor value as per data size*/
-std::optional<double> fetchSensorValue(const pldm_numeric_sensor_value_pdr& pdr,
-                                       const union_sensor_data_size& data);
+std::optional<float> fetchSensorValue(const pldm_numeric_sensor_value_pdr& pdr,
+                                      const union_sensor_data_size& data);
 
 /** @brief Get sensor unit as per D-Bus representation*/
 std::optional<SensorUnit>
     getSensorUnit(const pldm_numeric_sensor_value_pdr& pdr);
 
 /** @brief Fetch range field value as per range field format*/
-std::optional<double>
+std::optional<float>
     fetchRangeFieldValue(const pldm_numeric_sensor_value_pdr& pdr,
                          const union_range_field_format& data);
